read target long size for page size
The ring buffer page uses local_t to keep track of the size of data
held on an individual page. local_t is a long which means that it is
either 4 bytes on a 32bit box, or 8bytes on a 64bit box.

The code must be able to cope with this difference.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>