Fix bug that made %pf not work at the end of a fmt (and %pF not work at all)
When parsing %pf or %pF, pretty_print checked if the next character
matched 'f' or 'F', and if so, set show_func to the character *after*
that.  If not at the end of the string, this would still have the effect
of 'f', since pretty_print only checked for non-zero show_func; however,
'F' would not work.  If at the end of the string, show_func would get
the '\0', which disabled function printing entirely.

Signed-off-by: Josh Triplett <josh@joshtriplett.org>