allow plugins to override the event parsing handler
This patch adds the function to parse_events called

  pevent_register_event_handler(int id, char *sys_name,
			char *event_name,
			pevent_event_handler_func func)

If id is >= 0 then the event is searched for by id, and checked
against sys_name and event_name if those are non NULL.

If id < 0, then the sys_name and event_name are used. If sys_name
is non NULL, it will match the event system with that name,
otherwise it will just find the first event named event_name.

The pevent_event_handler_func is defined as:

  int (*func)(struct trace_seq *s, void *data, int size);

This lets a plugin register a function that will be called in place
of the "pretty_print" function for the event.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>