keep track of buffer full
The trace_seq buffer might fill up, and right now
one needs to check the return value of each printf
into the buffer to check for that.

Instead, have the buffer keep track of whether it
is full or not, and reject more input if it is full
or would have overflowed with an input that wasn't
added.

Also simplify the buffer output function and have
it print [truncated] when that is the case.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>