Move most read functions from trace-read.c to new trace-input.c
The trace-input.c file belongs to the libtracecmd library. Move the
functions from trace-read that will be used as an interface for
the library.

Note, this commit is a basic move. Only some changing from static
to global was done to some functions and variables to keep the code
working. But this is a transition commit. This shows the move.
The next commits will show the changes to these functions to make
the library API more practical.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>