Move reading of data into libtracecmd library
Moved the functions that deal with reading the trace file from
trace-read.c to trace-input.c. Tried to not modify as much as I could,
but I needed to convert some of the utils to use the library routines
as well as pass around the handler through function parameters.

Also added a hack to keep trace_peek_data that is used by libparsevent
for the function graph tracer.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>