Have tracecmd_read_at return a record that contains the offset
Currently if the offset requested is not the same as the offset for
the record, it will return the next record. We should instead return
the record that the offset is in.

TODO: known issue, the record->offset is the offset of the actual
record and not data. The record->size is the size of the data and not
the size of the actual record. A offset to the end of the record
may still return the next record. To fix this, we should add a
record_size.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>