Add translate_data to handle straight data translation to record
We add tracecmd_translate_data that will take straight data and
convert it to a record. Since raw data buids up its timestamp, it
will only be able to return the delta and not true timestamp.

This also consolidates this translation (although we don't handle
the old format with it).

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>