Prevent an infinite loop in searching by timestamp
If the search for a record comes down to being on a page
where the timestamp is greater than the page but less then the
next page to look for, the code will go into an infinite loop.
This is because the next page to look for will be the same as
the starting page, and it will never continue to move forward.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>