trace-cmd: Add sleep_time between splice reads
I found that tracing few events causes trace-cmd to almost spin
in a busy loop since it can constantly read a page, and the process
of reading that page causes it to run again. We end up just causing
events to read.

This adds a -s sleep time option that forces trace-cmd to sleep.
The default is 1ms. But it can be disabled with 0.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>