trace-cmd: Add initial python tracecmd API
Add an object oriented python API in a new tracecmd.py module. This
module contains a self test that duplicates the behavior of the
tracecmd-test.py, so it has been removed.

This API is far from complete and only implements enough functionality
to list the events on each CPU.

SWIG doesn't handle long longs so we have to add a wrapper, a
typemapper, and a marshaller to get a 64bit integer out of the record
data. This is a first take, comments welcome. We'll need to do something
similar with the timestamp values.

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>
LKML-Reference: <4B2AC3E1.1010301@us.ibm.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>