trace-cmd: Add ability to add new comms after first use
To speed up the reading of cmdlines to pid mappings, an unsorted list
is used first to register comms (cmdlines). Once the list is
used to seach, the list is converted to a sorted array to allow
for fast binary searches. But once this was done, there was no way
to add more cmdlines mappings.

This now adds a slower version to add new cmdlines after the list
has been made, but still makes it possible.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>