trace-graph: Handle function-graph hack
In order for the function graph tracer to know if the current
entry is a nested function or a leaf function, it must read ahead
to examine the next entry. But if the next entry happens to be on
another page, doing this will unmap the current record's data.

We have debug printing to the screen as we mouse over the events.
The debug prints out the record's info, and this can call the
function graph tracer's routine that will move the record.
But then we reference the record a second time to display the same
info in the drawing. This second one is using a unmapped page.

For now, we hack a work around by freeing the record and rereading
it after we display the debug info.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>