trace-cmd: Pass in record to pevent functions
Instead of passing in the data, size and timestamp to the
pevent parsing routines, pass in the record itself. This allows
the pevent parsing routines to have a bit more control and facilitates
the need of future work where we need to know more about the record
in the parser.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>