trace-cmd: Have function graph keep record around
The function graph tracer looks forward in the trace to determine
if a function is a leaf or not. In doing so, it causes the mapping
of records to change. This produces a nasty side effect where the
caller can have the record stored, and it will suddenly invalidate
the record's mapping.

This refreshes the record passed in and does a trick by calling
the tracecmd_peek_data function that will cache the location
after the leaf function's return entry.

This unfortunately causes a side effect too where if the last
entry in the trace is a leaf function, we can't hide the exit entry
and it will print regardless. But this side effect wont crash the
application, but the previous side effect will.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>