trace-cmd: Use file long size if no header page is found
Use the long size of the trace.dat file if no header page is
found, instead of using the long size of the host that is
running the report. It makes more sense to use the long size
that the record was running on than to read the host. This
way it at least breaks on all boxes and not just some.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>