trace-cmd: Add page references for records
It can be confusing for a user of tracecmd records to keep track
when its data is mapped. The data of a record points to a mmapped
page of the data file. But currently only one page is mapped at
a time. If two records are read that are on different pages, the
first record's memory will be unmmaped.

Instead of putting the management to the user, this patch changes
the design to keep a reference count to the page. When a record
is created, it adds one to the page's reference count. The page
is only unmmaped when the reference count goes to zero.

A check on tracecmd_close is made to make sure all pages have been
unmmaped, otherwise a warning is printed. This would happen if
a record is not freed properly.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>