trace-view: Self-adjusting fixed width trace-view
By using fixed width columns and fixed height rows, the view can now
render over 400k events without any noticable lag. This makes it so
large traces can now be viewed in one "page" of the trace-view-model.

The new trace-view-data-func() is called when the view wants to render
cell (which it only has to do when it's visible now that we use fixed
sizings). Here we check the string width in pixels and resize the column
if necessary. By tracking the max character width of each column, we can
avoid doing the expensive Pango pixel width test as often. By adding a
little extra padding, we anticipate one more character, further reducing
the overhead. The first full scroll down and up shows no noticable lag
due to these calculations once the optimizations were implemented.

V2: fix an unused var warning and use a long when casting to a pointer.

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>