trace-cmd: Add plugin_kmem for showing function names of call sites
The kmem events only show a useless hex symbol for the call site name.
This adds a plugin to do a lookup of the function name and still
print the rest of the event format.

Also adds pevent_find_function_address() to let the plugin show the
offset into the function as well.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>