trace-cmd: Have tracecmd_read_cpu_first() return the first record
There was a bug with tracecmd_read_cpu_first() where it may not return
the first record for a CPU. If another record was read on the first page
the get_page() function would return without reseting the index.

The tracecmd_read_cpu_first() needs to always reset the index before
returning tracecmd_read_data().

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>