trace-view: Fix searching by timestamp for filtered lists
When a list is filtered, the search_for_record_by_timestamp() was returning
incorrect values because the rows_ts_cmp() was using the pointer the array
was pointing to go to the next record. But if that record was not visible
or the next record should have been on another CPU, it would break the
bsearch.

The correct approach is to look at the next record in the array, not the
per cpu list.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>