trace-view: Fix trace_view_store_get_timestamp_visible_row()
The search used the size of the actual_rows array, and not the
size of what was visible.  The visible_rows use the array, but anything
outside the visible rows are just left overs.

Using the bsearch on the entire array causes errors in the search.
It should only use the subset that is visible.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>