tree-view: Have event filter list only expand rows not full or empty
The list of events for the event filters, now only expands the system
rows that are not full or empty. That is, if the system is checked
it is not expanded. If the system is not checked and has no events
checked, then it is not expanded.

But if the system is not checked, and there are events checked
(even all events) then the row is expanded.

This makes dealing with lots of events much easier.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>