trace-graph: Keep event info in view
If the scrolled window is vertically adjusted so that the top CPU
bar is close to the top of the viewable area, the event info box
can be hidden.

Change the location of the event info box to go beneath the pointer
when near the top.

The location of the box is also changed to be to the left of the
pointer instead of the right. This is because when we make the flip
the box may be hidden by the cursor itself. Although this does not
protect against that happening when the pointer is near the top
left corner, since the box will be to the right of the cursor then
anyway.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>