trace-graph: Allow selection before graph for zooming
It is difficult to zoom in from the beginning of the graph, because
you must move the pointer right to zoom. To ease zooming in at
the start of the graph, this patch now allows the user to select
on the CPU drawing, and move the pointer over the graph to make
the selection.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>