trace-cmd: Rename tracecmd_open to tracecmd_alloc
The tracecmd_open() currently returns a tracecmd_input descriptor
that is not finished. The headers still need to be read and data
needs to be initialized.

Rename tracecmd_open and tracecmd_open_fd to tracecmd_alloc and
tracecmd_alloc_fd respectively, and create a new tracecmd_open()
and tracecmd_open_fd() that returns a usable descriptor.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>