trace-cmd: Add sched_switch to and wake up of pid to filter
If the -F option is set, include events that wake up that pid or
switch to it. That is, even if the event is done by another task,
if that task calls the wakeup event to the filtered pid, it should
also be shown, as well as the context switches to the filtered task.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>