kernelshark: Create and use trace_view_store_get_visible_row()
The gtk_tree_selection_get_selected_rows returns the row of the
entire list, no the visible row. The current trace_view_store_get_row()
returns the record of the row passed in compared to the visible row.
But the row that is passed in is the absolute row, and this causes
us to index incorrectly.

This patch creates trace_view_store_get_visible_row() for use when
the row passed in is the visible row, and changes
trace_view_store_get_row() to return record based off of the
absolute row.

Reported-by: Darren Hart <dvhltc@us.ibm.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>