trace-graph: Start of graph_plot structure
This is a start of moving from plotting just the CPUs, but
to move the code to a plot interface. This will eventually
allow a line to just plot a task (on all CPUs).

This first patch just sets up some of the infratructure,
but still is tied to the CPU plots.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>