trace-view: Compare list all but last item in list
The last item is not added to the compare, because the compare
tests each item plus the next. If we compared the last item, it
may compare a invalid pointer.

But the code compares all but the last 2 items! Which means we
may return NULL when we should have found something. Compare that
second to last item.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>