trace-graph: Pass plot around instead of index
Use the plot itself to pass to routines instead of the index
into the ginfo->plot_array of the plot. This requires that the
plot keeps track of its own position in that array, and the
trace-plot.c code must update that position when plots are added
or removed from the middle of the array.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>