trace-graph: Limit the search back for task plots display data
The task plot was search every CPU for a matching event. If a CPU did
not have any events for a given plot, it would search all the way back to the
start!  This caused a huge delay in processing and would hang trace-graph
for seconds while it would search the cpu data backwards.

Limit the distance to only 5 pixels in its search.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>