trace-graph: Display task specific data when zoomed in on task plot
When zooming in on a task plot and the events show up in the graph,
instead of showing the CPU type data, show information useful for the task.
This includes what CPU it is on, and for sched_switch, if a wakeup
was detected, it shows the latency from the wakeup.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>