trace-cmd: Add locked attribute to record and debug peek freeing
Add a "locked" attribute to the record structure. This will be
set when a peek is returned and cleared on read. This will help
debug cases that incorrectly free a record returned by peek.

The internal helper function free_next() is added to do the freeing
of the next record used by peek.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>