parse-events: Make input buffer const char *
Make the input buffer that the parse events reads into a const char *.
This will be needed by later patches to make sure the input buffer
does not get modified.

In doing this change, the work around by Johannes Berg needed to
be rewritten, since the push_str modified the input buffer.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>