trace-cmd: Relicense the library files under LGPL
All the files that are used to create the libraries:
 libparsevent and libtracecmd are converted to the LGPL
as well as the files to create the plugins.

All files now have a boilerplate header that states which
license that the file is under. A README file is created as
well as the COPYING.LIB which contains the text of the
LGPL.

All authors of the code that created these files have given
their Acks.

Acked-by: Darren Hart <dvhltc@us.ibm.com>
Acked-by: Frederic Weisbecker <fweisbec@gmail.com>
Acked-by: Josh Triplett <josh@joshtriplett.org>
Acked-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>