parse-events: Change report filters to allow regex on event names
Use the regex to find events and systems. Since we are now using regular
expresions to add events to filters, we can no longer use the '.' to
separate the system name from the event name. That is:

 trace-cmd report -F 'sched.sched_switch'

will no longer work. It needs to be:

 trace-cmd report -F 'sched/sched_switch'

Which is a better naming since that is how the files are distinguished
anyhow.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>