parse-events: Cache last event found
Code commonly passes around the event id and uses that to find the
event. Even if the event was already found, it uses the id to find
it later.

By caching the last event found by the name or id, and checking that
on the next query before searching, saves some time.

Running trace-cmd report on 6,000,000 events went from:

 real	0m41.805s
 user	0m39.483s
 sys	0m2.311s

to:

 real	0m39.769s
 user	0m37.401s
 sys	0m2.356s

About a 4% speed-up.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>