trace-cmd: Add new trace-cmd listen, and record -N
Add a way to record the trace remotely. On a remote box, one
can start trace-cmd like:

   trace-cmd listen -p 1234

Then the local box that wants to trace can run:

   trace-cmd record -e all -N otherhost:1234 command

Instead of writing to trace.dat, the record will send the
data over the network using TCP and UDP and the trace.dat
file will be created on the remote host.

The remote host will save the file as trace.host:port.dat

TCP is used to send the recording box information, and UDP is
used to send the live data from the trace.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>