trace-cmd: Do not bail on failing mmap on second CPU if first had no size
If the first CPU had no size, and the second CPU failed to mmap, do
not fail the entire trace-cmd report. The mmap check is not done
on CPUs with no size.

If the mmap fails, then check the previous CPUs to see if any of
them had size, and if so then bail. But if none did, then
continue.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>