trace-view: Keep selected row after CPU filters
After a CPU filtering event happened, the selected row was lost
and the first row is selected. This gets annoying when you have
a row selected and want to filter out the other CPUs. You would
lose the row and then have to search for it again.

This patch reselects the selected row after a CPU filter change.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>