trace-cmd: Add Makefile to build man pages
Add a Makefile that has logic to build a man page.
It is still very primitive and uses ascriidoc and
xsltproc to do the formating. But instead of build my own xsl
file or putting together one from others, I decided to use
xsl files that already exist by doing a locate on
manpages/docbook.xsl, and using what is found.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>