parse-events: Print field contents of failed parsed event
If an event fails to parse the print string, then it is better to show the
contents as raw elements than just to print "FAILED TO PARSE" and drop the
event.

At least now the events show a bit more information, even if the print string
did not parse correctly.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>