trace-cmd: Clean up trace-cmd usage output
The usage just printed all commands and their full descriptions. This is too much
for all the commands that trace-cmd does.

This patch moves the usage out of trace-cmd.c and into a new file called trace-usage.c.
It then makes a structure that holds the name of the command, a short description
and the full usage (from the original usage).

If no command, or the command passed to trace-cmd is not found, then the list
of commands and their short usages are shown.

If a command is passed in, only the usage for that command is shown.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>