build: Prevent rebuild of documentation every time
Using dependencies to trigger the manpage.xsl waring was causing all man pages
to be built every time, event when no files were changed.

By converting it to a funciton and calling it first, it will allow the man pages
to only be rebuilt if their source files were changed.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>