trace-cmd: Make sure trace-cmd split start and end values are floats
It is a common mistake to do:

 trace-cmd split my-trace.dat

Expecting to be splitting the file my-trace.dat. But this is taken
as the starting value and the user will not know that the command
ran the split against trace.dat (if it exists).

This patch checks to make sure that the start and end values passed
in are indeed floating point (or integer) values.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>