trace-cmd: Remove extra page copy in trace-cmd split -p
The while loop that copies pages from one dat file ot the next
breaks after it copies the last page that it is expected to.
But after the while loop (for other split types) there is also a checkt
if a page is not finished and should be copied. But this always copies
the page again with the page count option, and causes a duplict page
at the end of the trace.

The fix is simply to break out of the loop before copying the data
when the page count is hit. Then the copy after the while loop will
copy the last page needed.

Reported-by: Li Zefan <lizf@cn.fujitsu.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>