parse-events: Fix parenthesis parsing of filters
The parenthesis filter parsing was a bit broken and needed a rewrite.
This redesigns the filter parsing to be a bit more sane when it comes
to parenthesis. It lets the parenthesis be treated as what they resolve
to be.

Funny, the example used by kernelshark failed because of this bug:

 "next_prio < 100 && (prev_prio > 100 && prev_pid != 0)"

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>