build/doc: Fix removing of trace.dat(5) SYNOPSIS
To remove the trace.dat man page "SYNOPSIS" section, the lines
from: .SH "SYNOPSIS" to .sp  are removed. But some systems do not
add a .sp after the SYNOPSIS, so they remove the entire rest of
the document.

Instead of testing for .sp, test for "ignore" since that is in the
document and is known to be there.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>