python wrapper: do not pass Trace but pevent
When it will be used for the python plugin,
an Event will not have a Trace pointer in
python directly available as that would not
make a lot of sense since it is used for
reading a trace file. But the event also
doesn't need a Trace pointer, it just needs
the pevent.

Acked-by: Darren Hart <dvhltc@us.ibm.com>
Signed-off-by: Johannes Berg <johannes@sipsolutions.net>