trace-cmd/split: Force percpu when parsing by pages is set
When parsing by pages is set, we automatically set the percpu parsing.
This is because non percpu does not make sense. The CPU data is stored
per page and two or more CPU data is not shared on a single CPU data page.

This patch forces the -c option (percpu) when the -p option is set.
Previously the -p option without the -c option gave some strange results.
Usually just one or two events per page.

Reported-by: Li Zefan <lizf@cn.fujitsu.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>