trace-cmd: make records refcounted
If a callback wants to hold on to a record,
it can then increase the refcount and keep
it. This is most likely to happen with the
upcoming python plugin.

Since the underlying data pages are already
refcounted, we can just add a refcount to
struct record (rather than having the error
checking with the 'locked' member).

Acked-by: Darren Hart <dvhltc@us.ibm.com>
Signed-off-by: Johannes Berg <johannes@sipsolutions.net>