trace-cmd: Copy field name in old string processing
In processing old print-fmt's (and also when the print format does not contain
args), the arg field name points directly to the field name. But on at the
end it will free the field name twice. One for the arg and one for the field.
The name needs to be strdup().

Reported-by: Lai Jiangshan <laijs@cn.fujitsu.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>