trace-cmd: unify plugin loading
Introduce a new externally visible function
trace_util_load_plugins that takes a callback
to load a given plugin, and use it in both
the core to load .so plugins, and the python
plugin to load .py plugins.

Note that this changes the location from which
python plugins are loaded.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>

[ cleaned up whitespace damage ]

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>