build/doc: Use a substitution reference in man pages install
subst will substitution every occurence of a string with another.
What we want here is just to substitute the last occurence.
That is we want to transform files that end with .1 into .1.install
We do not want to transform every ".1" in a path name into .1.install.

Without this fix you can get errors of this nature.

 make prefix=/home/jkacur/rpmbuild/BUILDROOT/trace-cmd-1.0.1-4.fc12.x86_64/usr install_doc
 make -C /home/jkacur/rpmbuild/BUILD/trace-cmd-1.0.1/Documentation install
 make[1]: *** No rule to make target `/home/jkacur/rpmbuild/BUILD/trace-cmd-1.0.1.install/Documentation/trace-cmd.1', needed by `/home/jkacur/rpmbuild/BUILD/trace-cmd-1.0.1.install/Documentation/trace-cmd.1.install'.  Stop.
 make: *** [install_doc] Error 2

Signed-off-by: John Kacur <jkacur@redhat.com>
LKML-Reference: <1276099288-3029-3-git-send-email-jkacur@redhat.com>

Use a "substitution reference" instead of a substitution.
A "subst" will change every occurence of a string
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>