kernelshark: Use pevent instead of handle for filter dialogs
The internal filter dialogs currently pass around the tracecmd_handle,
but only uses the pevent to display the data. Pass pevent around
instead. This also allows us to add:

 trace_filter_pevent_dialog()

This function will display a event dialog with a pevent being
passed to it instead of a tracecmd_input handle. Now we can
pop up the events local to the system that are not related to
any file.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>