trace-graph: Account for events not found in plot
In the plot code, it is possible that an event exists in the
data file that was not stored in the events log. This happened
with the 2.6.35 kernel where some events had formats that were "TOO BIG"
and that event format was not saved. But the events still existed in
the trace.dat file.

The plot code assumed that events were found and when they were not
it crashed.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>