trace-cmd: Always enabled tracing_enabled
The tracing_enabled file really needs to be deprecated. It should
not be used. But if the tracing_enabled file gets set to zero
somehow (by either the user or tool - latencytop?) then it will
disable output.

Since trace-cmd does not touch the file, it will confuse users
when they see no data recorded.

The simple answer is to always set tracing_enabled to one if the
file exists.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>