trace-cmd: Use correct size for rest of record
When the field does not have a lenth, use rest of the data item.
But account for the offset of the field.

Field arrays with [0] denote to use the rest of the record.
But the size set was the size of the record and not the size
starting from the field to the end  of the record.

This was causing SEGFAULTS when the record was at the end of
a page, and the copy was going beyond the page.

Reported-by: Will Schmidt <will_schmidt@vnet.ibm.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>