kernelshark: Have trace_create_combo_box() be able to just return combo box
Instead of requiring trace_create_combo_box() to work with a hbox,
change it so passing in a NULL hbox will just produce a combo box
and ignore the text (for the label).

This will allow us to still use this helper function when we need to
create the label and combo on something other than a hbox.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>