kernelshark: Change cancel button on capture dialog to close
"Cancel" is confusing and people may think that the Cancel will
just stop the trace and do nothing more. But this is incorrect.
The cancel button will stop the trace and update the data for the
graph and list.

Changing it to "Close" shows that it just closes the window.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>