trace-cmd: Add option to ignore event not found error
Currently, if 'trace-cmd start' passes an event that is not found
it will cause the command to fail.

This can be a problem for an administrator using trace-cmd on several
machines that have different modules configured in. One script may
be used that has all the events to be traced, but some of the events
may not exist on all machines. This means each machine will need
its own separate script.

Instead of having a separate script, add a '-i' option that lets one
ignore failed events.

Signed-off-by: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
LKML-Reference: <20100629144832.38C6.A69D9226@jp.fujitsu.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>