trace-cmd: Fix trace-cmd function graph handling of its events
The function graph needs to store the function graph return event
for later use. This patch changes the function graph callbacks
to use the context to store this information and initialize it
when first referenced.

The event data is now stored on the tracecmd_input handle, that allows
more than one tracecmd_input handle to work with this code in one
utility.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>