parse-events: Make trace_seq dynamically expand its buffers
The trace_seq code that is used to copy data into a buffer to later
be able to print was originally taken from the Linux kernel. But the
constraints on the kernel are different than what they are in userspace.
Since the kernel uses the trace_seq in critical paths, it can not
allocate memory during use, so it used a static page buffer.

But in userspace, we do not have this constraint, and are allowed to
realloc the buffer when needed. There's no reason to fill the buffer
when we hit the buffer size, we can simply allocate more memory.

The downside to this is that we must call a destructor on the trace_seq
structure. All trace_seq's must be initialized with trace_seq_init(),
so this patch matched all trace_seq_init()'s with the new function
trace_seq_destroy().

Suggested-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>