trace-graph: Fix plots to show start bars in zoomed views.
Currently, if a zoomed view has events before the start of the view,
those events are not examined to determine if the bar of a graph should
be shown at the start.

This patch adds code to the cpu and task plots to look for events
that happened before the view to determine if a bar should be shown
or not.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>