trace-cmd/listen: Force search of unique port
On some systems, the code to find a unique udp port may
give a invalid success of the previous port. Which causes two
child process to listen to the same port, and makes one of them
fail.

Force the start of the port search to the port after the last
port found, to prevent the same port from being used by two
listeners.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>