trace-cmd: Reset record pointer in set_cpu_to_timestamp()
The code to reset the cpu to a specific timestamp some time fails
if the time stamp does not match the current saved time stamp
and we go to another page. Without resetting the timestamp
for that page, the algorithm can get confused if the timestamp
is currently in the middle of the page and the record we
want is on the page but before the timestamp.

Simply reset the timestamp to the start of the page when
going to another page.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>