parse-events: Add PRINT_BPRINTK to separate bprintk strings from dynamic
When bprintk had a "%s",string format, the args for the bprintk used
a dynamic string arg for the string. Unfortunately, this would cause
the string to not be found and nothing would be printed.

This patch adds a special PRINT_BPRINTK type that lets the processing
do something special with the bprintk strings.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>