parse-events: Add filter on task comm
Add a 'COMM' special field to allow the filter in trace-cmd report
to filter on the task's comm.

Most records do not have data about what the comm is of the
task that recorded the record. But comms are stored in various
ways that parse-events can handle.

By adding a special field 'COMM' (all caps) the user can now
filter out tasks based on their name. This is useful when filtering
out (or in) a bunch of threads.

  -F 'COMM != "trace-cmd"'

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>