parse-events: Use arg field size to test for pointer
Using the "len" variable in print_str_arg() to test if the
value is an array when the arg field size is zero is incorrect.
When the arg field size is zero, the data is an dynamic value
regardless if FIELD_IS_ARRAY is set or not. It may not be set
for strings at the end of the record.

Only assume the data is a pointer if the arg->field.field->size
is sizeof long.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>