trace-cmd: Add events to handle sched switch/wakeup filters
The sched_switch and sched_wakeup filters are special for tracing
pids. If we are tracing a pid, we want sched_switch and sched_wakeup
to be traced if it is waking up or switching to the given pid.

Since events are now expanded into a list holding the filter
files to be updated. Either reuse or create if necessary the
sched_* events to add the extra filters.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>